+++ scripts/screens/SettingsScreen.py
@@ -22,9 +22,9 @@ from scripts.game_structure.game.settings import (
 import scripts.game_structure.game.settings.settings as all_settings
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
-    UIImageHorizontalSlider,
     UICheckbox,
 )
+from ..ui.elements.image_horizontal_slider import UIImageHorizontalSlider
 from ..ui.elements.modified_scrolling_container import UIModifiedScrollingContainer
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
